modificacion de la migracion de consultation_prices (t.integer :amount)

+++ db/migrate/20100909193719_create_consultation_prices.rb
@@ -2,7 +2,7 @@ class CreateConsultationPrices < ActiveRecord::Migration
   def self.up
     create_table :consultation_prices do |t|
       t.string :price_type
-      t.decimal :amount
+      t.integer :amount
 
       t.timestamps
     end
